requirements.txt
torch==2.0.1
torchvision==0.15.2
layoutparser
opencv-python
Pillow
numpy==1.26.4